# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: User not found in context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Requested file not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: Internal server error (S3, DB, or parse failure)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
    OptId:
      name: opt_id
      in: query
      required: true
      schema:
        type: string
      example: MH_WMIT_SN_NS046496
...
paths:
...
  /get_state_district:
    get:
      summary: List all distinct states and districts
      operationId: GetStateDistrict
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  states: { type: array, items: { type: string } }
                  districts: { type: array, items: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /selected_registrars:
    post:
      summary: Get risk distribution for a chosen set of registrars
      operationId: GetSelectedRegistrars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selected_registrars]
              properties:
                selected_registrars:
                  type: array
                  items: { type: string }
                  minItems: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  reg_distribution:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        high_risk: { type: integer }
                        med_risk: { type: integer }
                        low_risk: { type: integer }
                        no_risk: { type: integer }
        "400":
          description: Invalid body or empty selected_registrars
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /all_registrars:
    get:
      summary: Get all registrar names for caller's RO (sorted)
      operationId: GetAllRegistrars
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  registrars: { type: array, items: { type: string } }
                  count: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /operator_details:
    get:
      summary: Get an operator's opt_details.json
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted.
      operationId: GetOperatorDetails
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file: { type: string }
                  data: { type: object, additionalProperties: true }
                  requested_by: { type: string }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or file not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /operator_risk_details:
    get:
      summary: Get an operator's risk_details.json
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted.
      operationId: GetOperatorRiskDetails
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file: { type: string }
                  data: { type: object, additionalProperties: true }
                  requested_by: { type: string }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or file not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /region_evaluation_count:
    get:
      summary: Get audit.json evaluation counts for an RO/state/district
      operationId: GetRegionEvaluationCount
      parameters:
        - name: regional_office
          in: query
          required: true
          schema: { type: string }
          example: Mumbai
        - name: opt_state
          in: query
          required: false
          schema: { type: string }
        - name: opt_district
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  opt_state: { type: string }
                  opt_district: { type: string }
                  file: { type: string }
                  data: { type: object, additionalProperties: true }
                  requested_by: { type: string }
        "400":
          description: regional_office missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...